# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Book Store API Documentation'
  description: 'This API for educational purpose with study case book store. Created by Satya Wibawa.'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
paths:
...
  /api/v1/tags:
    get:
      summary: 'Tag Collection'
      operationId: tagCollection
      description: 'Get tag collection'
      parameters:
        -
          in: query
          name: q
          description: 'Searching keyword of the collection.'
          example: Encyclopedia
          required: false
          schema:
            type: string
            description: 'Searching keyword of the collection.'
            example: Encyclopedia
        -
          in: query
          name: pageSize
          description: 'Count of items per page. Allowed size 10, 25, 50, 100. Default 10.'
          example: 10
          required: false
          schema:
            type: integer
            description: 'Count of items per page. Allowed size 10, 25, 50, 100. Default 10.'
            example: 10
        -
          in: query
          name: page
          description: 'Page index to view.'
          example: 1
          required: false
          schema:
            type: integer
            description: 'Page index to view.'
            example: 1
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  tags:
                    -
                      id: 1
                      name: Dictionary
                    -
                      id: 1
                      name: Dictionary
                  message: 'Success to get tag collection'
                properties:
                  tags:
                    type: array
                    example:
                      -
                        id: 1
                        name: Dictionary
                      -
                        id: 1
                        name: Dictionary
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        name:
                          type: string
                          example: Dictionary
                  message:
                    type: string
                    example: 'Success to get tag collection'
      tags:
        - 'Tag Management'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                pageSize:
                  type: integer
                  description: 'Count of items per page. Allowed size 10, 25, 50, 100. Default 10.'
                  example: 10
                q:
                  type: string
                  description: 'Searching keyword of the collection.'
                  example: Encyclopedia
      security: []
    post:
      summary: 'Store new tag'
      operationId: storeNewTag
      description: 'Store a newly created tag in database.'
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        201:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  tag:
                    id: 2
                    name: Magazine
                  message: 'Tag has been created.'
                properties:
                  tag:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 2
                      name:
                        type: string
                        example: Magazine
                  message:
                    type: string
                    example: 'Tag has been created.'
      tags:
        - 'Tag Management'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: "Tag's name."
                  example: Magazine
              required:
                - name
      security: []
...